issue #42 (refactor - test): refactoring test/util/util.js + using it in topics integration tests
